improve slippage custom input
