+++ src/components/swapv2/SwapSettingsPanel/SlippageSetting.tsx
@@ -1,164 +1,14 @@
 import { Trans, t } from '@lingui/macro'
-import React, { useRef, useState } from 'react'
+import React from 'react'
 import { isMobile } from 'react-device-detect'
 import { Flex, Text } from 'rebass'
-import styled, { css } from 'styled-components'
+import styled from 'styled-components'
 
 import QuestionHelper from 'components/QuestionHelper'
-import { MAX_SLIPPAGE_IN_BIPS } from 'constants/index'
-import useMixpanel, { MIXPANEL_TYPE } from 'hooks/useMixpanel'
+import SlippageControl from 'components/swapv2/SlippageControl'
 import useTheme from 'hooks/useTheme'
 import { useUserSlippageTolerance } from 'state/user/hooks'
-
-const DefaultSlippages = [5, 10, 50, 100]
-
-const parseSlippageInput = (str: string): number => Math.round(Number.parseFloat(str) * 100)
-
-// isValid = true means it's OK to process with the number with an extra parse
-// isValid = true with message means warning
-// isValid = false with/without message means error
-const validateSlippageInput = (str: string): { isValid: boolean; message?: string } => {
-  if (str === '') {
-    return {
-      isValid: true,
-    }
-  }
-
-  const numberRegex = /^\s*([0-9]+)(\.\d+)?\s*$/
-  if (!str.match(numberRegex)) {
-    return {
-      isValid: false,
-      message: t`Enter a valid slippage percentage`,
-    }
-  }
-
-  const rawSlippage = parseSlippageInput(str)
-  if (Number.isNaN(rawSlippage)) {
-    return {
-      isValid: false,
-      message: t`Enter a valid slippage percentage`,
-    }
-  }
-
-  if (rawSlippage < 0) {
-    return {
-      isValid: false,
-      message: t`Enter a valid slippage percentage`,
-    }
-  } else if (rawSlippage < 50) {
-    return {
-      isValid: true,
-      message: t`Your transaction may fail`,
-    }
-  } else if (rawSlippage > MAX_SLIPPAGE_IN_BIPS) {
-    return {
-      isValid: false,
-      message: t`Enter a smaller slippage percentage`,
-    }
-  } else if (rawSlippage > 500) {
-    return {
-      isValid: true,
-      message: t`Your transaction may be frontrun`,
-    }
-  }
-
-  return {
-    isValid: true,
-  }
-}
-
-const EmojiContainer = styled.span`
-  flex: 0 0 12px;
-  ${({ theme }) => theme.mediaWidth.upToSmall`
-        display: none;
-  `}
-`
-
-const SlippageOptionCSS = css`
-  flex: 0 0 24%;
-  height: 100%;
-  padding: 0;
-  border: 1px solid transparent;
-  border-radius: 20px;
-
-  background-color: ${({ theme }) => theme.tabBackgound};
-  color: ${({ theme }) => theme.subText};
-  text-align: center;
-
-  font-size: 12px;
-  font-weight: 400;
-  line-height: 16px;
-
-  outline: none;
-  cursor: pointer;
-
-  :hover {
-    border: 1px solid ${({ theme }) => theme.bg4};
-  }
-  :focus {
-    border: 1px solid ${({ theme }) => theme.primary};
-  }
-
-  &[data-active='true'] {
-    background-color: ${({ theme }) => theme.tabActive};
-    color: ${({ theme }) => theme.text};
-
-    font-weight: 500;
-  }
-`
-const DefaultSlippageOption = styled.button`
-  ${SlippageOptionCSS}
-  flex: 0 0 18%;
-  @media only screen and (max-width: 375px) {
-    font-size: 10px;
-    flex: 0 0 15%;
-  }
-`
-
-const CustomSlippageOption = styled.div`
-  ${SlippageOptionCSS}
-
-  display: inline-flex;
-  align-items: center;
-  padding: 0 4px;
-  column-gap: 2px;
-  flex: 1;
-  input {
-    width: 100%;
-    height: 100%;
-    border: 0px;
-    border-radius: inherit;
-
-    color: inherit;
-    background: transparent;
-    outline: none;
-    text-align: right;
-
-    &::-webkit-outer-spin-button,
-    &::-webkit-inner-spin-button {
-      -webkit-appearance: none;
-    }
-  }
-
-  &[data-active='true'] {
-    color: ${({ theme }) => theme.text};
-    font-weight: 500;
-  }
-
-  &[data-warning='true'] {
-    border: 1px solid;
-    border-color: ${({ theme }) => theme.warning};
-
-    ${EmojiContainer} {
-      color: ${({ theme }) => theme.warning};
-    }
-  }
-
-  &[data-error='true'] {
-    border: 1px solid;
-    border-color: ${({ theme }) => theme.red1};
-  }
-`
+import { checkRangeSlippage } from 'utils/slippage'
 
 const Message = styled.div`
   font-size: 12px;
@@ -174,69 +24,12 @@ const Message = styled.div`
   }
 `
 
-const CustomInput = styled.input`
-  ::placeholder {
-    font-size: 12px;
-  }
-  @media only screen and (max-width: 375px) {
-    font-size: 10px;
-  }
-`
-
-const getSlippageText = (rawSlippage: number) => {
-  const isCustom = !DefaultSlippages.includes(rawSlippage)
-  if (!isCustom) {
-    return ''
-  }
-
-  if (rawSlippage % 100 === 0) {
-    return String(rawSlippage / 100)
-  }
-
-  return (rawSlippage / 100).toFixed(2)
-}
-
 const SlippageSetting: React.FC = () => {
   const theme = useTheme()
-  const inputRef = useRef<HTMLInputElement>(null)
-
-  // rawSlippage = 10
-  // slippage = 10 / 10_000 = 0.001 = 0.1%
-  const [rawSlippage, setRawSlippage] = useUserSlippageTolerance()
-
-  const isCustomOptionActive = !DefaultSlippages.includes(rawSlippage)
-  const [slippageInput, setSlippageInput] = useState(getSlippageText(rawSlippage))
-  const { isValid, message } = validateSlippageInput(slippageInput)
-
+  const [rawSlippage] = useUserSlippageTolerance()
+  const { isValid, message } = checkRangeSlippage(rawSlippage)
   const isWarning = isValid && !!message
   const isError = !isValid
-  const { mixpanelHandler } = useMixpanel()
-  const handleCommitChange = () => {
-    if (!isValid) {
-      return
-    }
-
-    if (slippageInput === '') {
-      setSlippageInput(getSlippageText(rawSlippage))
-      return
-    }
-
-    const newRawSlippage = parseSlippageInput(slippageInput)
-    if (Number.isNaN(newRawSlippage)) {
-      return
-    }
-    mixpanelHandler(MIXPANEL_TYPE.SLIPPAGE_CHANGED, { new_slippage: newRawSlippage / 100 })
-
-    setRawSlippage(newRawSlippage)
-    setSlippageInput(getSlippageText(newRawSlippage))
-  }
-
-  const handleKeyUp = (e: React.KeyboardEvent<HTMLInputElement>) => {
-    if (e.key === 'Enter') {
-      handleCommitChange()
-      inputRef.current?.blur()
-    }
-  }
 
   return (
     <Flex
@@ -265,61 +58,7 @@ const SlippageSetting: React.FC = () => {
         />
       </Flex>
 
-      <Flex
-        sx={{
-          justifyContent: 'space-between',
-          width: '100%',
-          maxWidth: '100%',
-          height: '28px',
-          borderRadius: '20px',
-          background: theme.tabBackgound,
-          padding: '2px',
-        }}
-      >
-        {DefaultSlippages.map(slp => (
-          <DefaultSlippageOption
-            key={slp}
-            onClick={() => {
-              setSlippageInput('')
-              setRawSlippage(slp)
-              mixpanelHandler(MIXPANEL_TYPE.SLIPPAGE_CHANGED, { new_slippage: slp / 100 })
-            }}
-            data-active={rawSlippage === slp}
-          >
-            {slp / 100}%
-          </DefaultSlippageOption>
-        ))}
-
-        <CustomSlippageOption
-          data-active={isCustomOptionActive}
-          data-warning={isCustomOptionActive && isWarning}
-          data-error={isCustomOptionActive && isError}
-        >
-          {isCustomOptionActive && isWarning && (
-            <EmojiContainer>
-              <span role="img" aria-label="warning">
-                ⚠️
-              </span>
-            </EmojiContainer>
-          )}
-          <CustomInput
-            ref={inputRef}
-            placeholder={t`Custom`}
-            value={slippageInput}
-            onChange={e => setSlippageInput(e.target.value)}
-            onBlur={handleCommitChange}
-            onKeyUp={handleKeyUp}
-          />
-          <Text
-            as="span"
-            sx={{
-              flex: '0 0 12px',
-            }}
-          >
-            %
-          </Text>
-        </CustomSlippageOption>
-      </Flex>
+      <SlippageControl />
 
       {!!message && (
         <Message data-warning={isWarning} data-error={isError}>
