+++ src/constants/index.ts
@@ -251,6 +251,8 @@ export const EPSILON = 0.000000000008854
 
 // https://www.nasdaq.com/glossary/b/bip
 export const MAX_SLIPPAGE_IN_BIPS = 2000
+export const DEFAULT_SLIPPAGES = [5, 10, 50, 100]
+export const DEFAULT_SLIPPAGE = 50
 
 export const AGGREGATOR_WAITING_TIME = 1700 // 1700 means that we at least show '.' '..' '...' '.' '..' '...'
